refactor: name convension and remove unused things
