+++ .env.example
@@ -9,4 +9,4 @@ APP_PATH=<LOCAL_PATH/App.app>
 ALLURE_RESULTS_DIR=tests/.reports/allure-results
 
 ## Choose the environment in the configs folder
-ENVIRONMENT=development #production
+ENVIRONMENT=dev-comm #dev-ent
