+++ package.json
@@ -103,7 +103,6 @@
     "jest-canvas-mock": "^2.5.2",
     "jest-environment-jsdom": "^29.5.0",
     "jest-transform-yaml": "^1.1.2",
-    "js-yaml": "^4.1.0",
     "lint-staged": "^14.0.1",
     "mini-css-extract-plugin": "^2.6.1",
     "node-polyfill-webpack-plugin": "^2.0.1",
