+++ src/core/configuration/configurationService.types.ts
@@ -1,0 +1,22 @@
+enum WitnessMode {
+  POOLS = "pools",
+  BACKER = "backer",
+}
+
+interface Configuration {
+  keri:
+    | {
+        witness: WitnessMode.POOLS;
+        pools: string[];
+      }
+    | {
+        witness: WitnessMode.BACKER;
+        backer: {
+          aid: string;
+          address: string;
+        };
+      };
+}
+
+export type { Configuration };
+export { WitnessMode };
