+++ src/core/configuration/index.ts
@@ -1,1 +1,1 @@
-export * from "./configuration";
+export * from "./configurationService";
